refactor: Change project structure to default (Standard Golang project layout)
